+++ benchmarks/generate_comparison.py
@@ -17,7 +17,7 @@ import matplotlib.pylab as plt
 import pandas as pd
 
 from benchmarks.test_basic_parity import measure_loops
-from tests.base.models import ParityModuleMNIST, ParityModuleRNN
+from tests.helpers.models import ParityModuleMNIST, ParityModuleRNN
 
 NUM_EPOCHS = 20
 NUM_RUNS = 50
