+++ tests/core/test_datamodules.py
@@ -24,7 +24,7 @@ from pytorch_lightning.accelerators.legacy.gpu_accelerator import GPUAccelerator
 from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.trainer.states import TrainerState
 from tests.base import BoringDataModule, BoringModel
-from tests.base.develop_utils import reset_seed
+from tests.helpers.utils import reset_seed
 
 
 def test_can_prepare_data(tmpdir):
