+++ tests/helpers/pipelines.py
@@ -17,7 +17,7 @@ from pytorch_lightning import Trainer
 from pytorch_lightning.trainer.states import TrainerState
 from pytorch_lightning.utilities import DistributedType
 from tests.base import BoringModel
-from tests.base.develop_utils import get_default_logger, load_model_from_checkpoint, reset_seed
+from tests.helpers.utils import get_default_logger, load_model_from_checkpoint, reset_seed
 
 
 def run_model_test_without_loggers(trainer_options, model, min_acc: float = 0.50):
