+++ tests/models/data/horovod/train_default_model.py
@@ -37,8 +37,8 @@ else:
     print('You requested to import Horovod which is missing or not supported for your OS.')
 
 from tests.base import EvalModelTemplate  # noqa: E402
-from tests.base.develop_pipelines import run_prediction  # noqa: E402
-from tests.base.develop_utils import reset_seed, set_random_master_port  # noqa: E402
+from tests.helpers.pipelines import run_prediction  # noqa: E402
+from tests.helpers.utils import reset_seed, set_random_master_port  # noqa: E402
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--trainer-options', required=True)
