+++ tests/models/test_onnx.py
@@ -18,8 +18,8 @@ import onnxruntime
 import pytest
 import torch
 
-import tests.base.develop_pipelines as tpipes
-import tests.base.develop_utils as tutils
+import tests.helpers.pipelines as tpipes
+import tests.helpers.utils as tutils
 from pytorch_lightning import Trainer
 from tests.base import BoringModel, EvalModelTemplate
