+++ tests/models/test_sync_batchnorm.py
@@ -20,8 +20,8 @@ from pytorch_lightning import LightningModule, seed_everything, Trainer
 from pytorch_lightning.plugins.legacy.ddp_plugin import DDPPlugin
 from pytorch_lightning.trainer.states import TrainerState
 from pytorch_lightning.utilities import FLOAT16_EPSILON
-from tests.base.develop_utils import set_random_master_port
 from tests.helpers.datamodules import MNISTDataModule
+from tests.helpers.utils import set_random_master_port
 
 
 class SyncBNModule(LightningModule):
