+++ tests/trainer/test_config_validator.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 import pytest
 
-import tests.base.develop_utils as tutils
+import tests.helpers.utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.base import EvalModelTemplate
