+++ tests/trainer/test_trainer.py
@@ -28,7 +28,7 @@ import torch
 from omegaconf import OmegaConf
 from torch.utils.data import DataLoader
 
-import tests.base.develop_utils as tutils
+import tests.helpers.utils as tutils
 from pytorch_lightning import Callback, LightningDataModule, LightningModule, Trainer
 from pytorch_lightning.callbacks import EarlyStopping, ModelCheckpoint
 from pytorch_lightning.core.saving import load_hparams_from_tags_csv, load_hparams_from_yaml, save_hparams_to_tags_csv
